Don't rebuild when our bot pushes new modularizations
Ultimately it would be better if the modularization script could
itself ensure that modularized boost was cloned, and at the right
revision (see what buildbot does to achieve this)... then I could
remove the repository from the bot's configuration altogether and go
back to using the default ChangeFilter.  For now, though, you need
something like a valid working copy of modularized Boost in order to
use the script.